updated servers port datatype.wip

+++ slsReceiverSoftware/src/File.h
@@ -58,7 +58,7 @@ class File : private virtual slsDetectorDefs {
     virtual void CreateFirstHDF5DataFile(
         const std::string &fileNamePrefix, const uint64_t fileIndex,
         const bool overWriteEnable, const bool silentMode,
-        const uint32_t udpPortNumber, const uint32_t maxFramesPerFile,
+        const uint16_t udpPortNumber, const uint32_t maxFramesPerFile,
         const uint64_t numImages, const uint32_t nPixelsX,
         const uint32_t nPixelsY, const uint32_t dynamicRange) {
         LOG(logERROR)
@@ -70,7 +70,7 @@ class File : private virtual slsDetectorDefs {
                                            const uint64_t fileIndex,
                                            const bool overWriteEnable,
                                            const bool silentMode,
-                                           const uint32_t udpPortNumber,
+                                           const uint16_t udpPortNumber,
                                            const uint32_t maxFramesPerFile) {
         LOG(logERROR)
             << "This is a generic function CreateFirstBinaryDataFile that "
